Update integration tests to use new API for mouse events
